+++ src/App.jsx
@@ -4,6 +4,8 @@ import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import ComingSoon from './ComingSoon';
 import Home from './Home';
 import Contact from './Contact';
+import Blog from './Blog';
+import PrivacyPolicy from './Privacy';
 import './index.css';
 
 function App() {
@@ -13,6 +15,8 @@ function App() {
         <Route path="/" element={<ComingSoon />} />
         <Route path="/details" element={<Home />} />
         <Route path="/contact" element={<Contact />} />
+        <Route path="/blog" element={<Blog />} />
+        <Route path="/privacy" element={<PrivacyPolicy />} />
       </Routes>
     </BrowserRouter>
   );